requirements.txt
pytesseract
pillow
langchain
openai
pydantic
langchain-community
langchain-openai
PyPDF2
pdf2image